# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Contract Risk Analyzer API
  description: |
    AI-powered contract analysis and negotiation assistance API.
    
    This API provides comprehensive contract analysis capabilities including:
    - Risk assessment and clause identification
    - Intelligent redline suggestions
    - Automated email generation for negotiations
    - Real-time progress tracking via WebSocket
    - Comprehensive health monitoring
    
    ## Authentication
    Currently, the API does not require authentication for basic operations.
    Future versions will implement JWT-based authentication.
    
    ## Rate Limiting
    - Synchronous Analysis: 10 requests per minute per IP
    - Asynchronous Analysis: 5 requests per minute per IP
    - WebSocket Connections: 3 concurrent connections per IP
    
    ## File Upload Requirements
    - Supported formats: PDF (.pdf), DOCX (.docx)
    - Maximum size: 10MB per file
    - Must contain readable text content
  version: 0.1.0
  contact:
    name: Contract Risk Analyzer Team
    url: https://github.com/your-org/contract-analyzer
    email: support@contract-analyzer.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.contract-analyzer.com
    description: Production server
paths:
...
  /api/v1/analyze-contract:
    post:
      tags:
        - Contract Analysis
      summary: Analyze contract synchronously
      description: |
        Analyze a contract document for risky clauses and generate negotiation suggestions.
        
        This endpoint provides immediate analysis results for smaller contracts.
        For larger documents, use the asynchronous endpoint.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Contract file to analyze (.pdf or .docx)
              required:
                - file
      responses:
        '200':
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
              example:
                risky_clauses:
                  - clause_text: "The Company shall not be liable for any indirect damages..."
                    risk_explanation: "This clause limits liability for indirect damages which could expose the client to significant financial risk."
                    risk_level: "High"
                    precedent_reference: "See similar clause in Contract ABC-123"
                    clause_index: 15
                suggested_redlines:
                  - original_clause: "The Company shall not be liable for any indirect damages..."
                    suggested_redline: "The Company shall not be liable for any indirect damages except those arising from gross negligence or willful misconduct..."
                    risk_explanation: "Original clause is too broad and could limit legitimate claims"
                    clause_index: 15
                    change_rationale: "Add exception for gross negligence to protect client interests"
                    risk_mitigated: true
                email_draft: "Dear [Counterparty],\n\nI've reviewed the contract and identified several clauses that require attention..."
                processing_time: 45.2
                analysis_timestamp: "2024-01-01T12:00:00Z"
                status: "completed"
                overall_risk_score: 7.5
                warnings: []
                errors: []
        '400':
          description: Invalid file or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '408':
          description: Analysis timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analyze-contract/async:
    post:
      tags:
        - Contract Analysis
      summary: Start asynchronous contract analysis
      description: |
        Start asynchronous contract analysis with progress tracking.
        
        This endpoint is recommended for larger documents or when you need
        real-time progress updates via WebSocket.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Contract file to analyze (.pdf or .docx)
                timeout_seconds:
                  type: integer
                  minimum: 30
                  maximum: 3600
                  default: 300
                  description: Analysis timeout in seconds
                priority:
                  type: string
                  enum: [low, normal, high]
                  default: normal
                  description: Analysis priority
                callback_url:
                  type: string
                  format: uri
                  description: URL to call when analysis is complete
                enable_progress_tracking:
                  type: boolean
                  default: true
                  description: Enable real-time progress tracking
              required:
                - file
      responses:
        '202':
          description: Analysis task started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncAnalysisResponse'
              example:
                task_id: "task_123456789"
                status: "pending"
                estimated_completion_time: "2024-01-01T12:05:00Z"
                status_url: "/api/v1/analyze-contract/async/task_123456789/status"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many concurrent tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to start analysis task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analyze-contract/async/{task_id}/status:
    get:
      tags:
        - Contract Analysis
      summary: Get analysis task status
      description: |
        Get the current status of an asynchronous analysis task.
        Includes progress updates and resource metrics.
      parameters:
        - name: task_id
          in: path
          required: true
          description: The ID of the analysis task
          schema:
            type: string
      responses:
        '200':
          description: Task status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisStatusResponse'
              example:
                status: "running"
                current_node: "risk_analysis"
                execution_id: "task_123456789"
                error_count: 0
                warnings: []
                last_error: null
                start_time: "2024-01-01T12:00:00Z"
                end_time: null
                processing_duration: 120.5
                risky_clauses_count: 3
                redlines_count: 2
                overall_risk_score: null
                contract_filename: "contract.pdf"
                progress_updates:
                  - node: "document_processing"
                    progress_percentage: 25.0
                    message: "Extracting text from document..."
                    timestamp: "2024-01-01T12:00:30Z"
                    estimated_remaining_seconds: 180
                resource_usage:
                  memory_mb: 512
                  cpu_percent: 45.2
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error retrieving task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analyze-contract/async/{task_id}/result:
    get:
      tags:
        - Contract Analysis
      summary: Get analysis task result
      description: |
        Get the result of a completed asynchronous analysis task.
        Only available for completed tasks.
      parameters:
        - name: task_id
          in: path
          required: true
          description: The ID of the analysis task
          schema:
            type: string
      responses:
        '200':
          description: Result retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '400':
          description: Task not completed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error retrieving result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analyze-contract/async/{task_id}/progress:
    get:
      tags:
        - Contract Analysis
      summary: Get task progress updates
      description: |
        Get progress updates for a specific task.
        Returns recent progress updates with timestamps.
      parameters:
        - name: task_id
          in: path
          required: true
          description: The ID of the analysis task
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number of progress updates to return
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Progress updates retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProgressUpdate'
              example:
                - node: "document_processing"
                  progress_percentage: 25.0
                  message: "Extracting text from document..."
                  timestamp: "2024-01-01T12:00:30Z"
                  estimated_remaining_seconds: 180
                - node: "risk_analysis"
                  progress_percentage: 75.0
                  message: "Analyzing clauses for risk factors..."
                  timestamp: "2024-01-01T12:02:15Z"
                  estimated_remaining_seconds: 45
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/analyze-contract/async/{task_id}:
    delete:
      tags:
        - Contract Analysis
      summary: Cancel analysis task
      description: |
        Cancel an active asynchronous analysis task.
        Only pending or running tasks can be cancelled.
      parameters:
        - name: task_id
          in: path
          required: true
          description: The ID of the analysis task to cancel
          schema:
            type: string
      responses:
        '200':
          description: Task cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Analysis task task_123456789 has been cancelled"
        '404':
          description: Task not found or cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error cancelling task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    AnalysisResponse:
      type: object
      required:
        - risky_clauses
        - suggested_redlines
        - email_draft
        - analysis_timestamp
        - status
        - warnings
        - errors
      properties:
        risky_clauses:
          type: array
          items:
            $ref: '#/components/schemas/RiskyClause'
          description: List of identified risky clauses
        suggested_redlines:
          type: array
          items:
            $ref: '#/components/schemas/RedlineSuggestion'
          description: List of suggested redlines
        email_draft:
          type: string
          description: Draft email for negotiation communication
          example: "Dear [Counterparty],\n\nI've reviewed the contract and identified several clauses that require attention..."
        processing_time:
          type: number
          nullable: true
          description: Time taken to process the contract in seconds
          example: 45.2
        analysis_timestamp:
          type: string
          format: date-time
          description: When the analysis was completed
          example: "2024-01-01T12:00:00Z"
        status:
          type: string
          description: Analysis status
          example: "completed"
        overall_risk_score:
          type: number
          nullable: true
          minimum: 0
          maximum: 10
          description: Overall risk score for the contract
          example: 7.5
        warnings:
          type: array
          items:
            type: string
          description: Processing warnings
          example: []
        errors:
          type: array
          items:
            type: string
          description: Processing errors
          example: []
...
  securitySchemes:
    # Future authentication schemes
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-based authentication (planned for future versions)
...